move ct utils file
